'mpo collect' now posts to api server:
EG of creating a new collection. This uses db.echo to send the post back.
[] mpo --format=pretty collect -n Gyro_runs -d "My first collection" -e 6f4a08de-0f3d-4364-9dfb-6e1237d98e9d
{
    "table":"collection",
    "data":"{\"elements\": [\"6f4a08de-0f3d-4364-9dfb-6e1237d98e9d\"], \"name\": \"Gyro_runs\", \"parent_uid\": null, \"description\": \"My first collection\"}"

EG of adding to a collection. This uses dryrun to show what the post would be:
[] mpo --dryrun --format=text collect -c 092966e4-8913-4ad1-a3a0-75429213adb7 -n Gyro_runs -d "My first collection" -e 6f4a08de-0f3d-4364-9dfb-6e1237d98e9d

DRYRUN in MPO.POST
MPO.POST to https://192.168.56.10:8443/v0/collection/092966e4-8913-4ad1-a3a0-75429213adb7/element with workflow:None, parent:092966e4-8913-4ad1-a3a0-75429213adb7 and payload of {"elements": ["6f4a08de-0f3d-4364-9dfb-6e1237d98e9d"], "name": "Gyro_runs", "parent_uid": "092966e4-8913-4ad1-a3a0-75429213adb7", "description": "My first collection"}


Note the collections present a new datastructure. In the past, adding an object to a workflow, or a comment/metadata to an object involved making a connection, hence the concept of a parent id in all posts. In the case of collections we are not making connections, just putting an object in a bag, so parent_uid is not a connection but is actually unnecessary as we post to the resource directly : /collection/<cid>/element when adding an element. So the 'parent_uid' is in the route. For consistency, I do assign parent_uid the same value as the <cid> in the route, but it is not needed. Note that when creating a new collection the parent_uid is null and the route is just /collection