Support for three api servers: mpoapi, test-api, and demo-api with routes starting with /api, /test-api, /demo-api
Which database you get is set in api_server.wsgi based on the process-group
name in the apache2-site-config (usually put in default-ssl under /etc/apache2)

web-server.wsgi still defaults to using /api. There should be a UI option, presently
needing to be hardcoded, to select between /api, /test-api, /demo-api as the api uri root.
Likely, session variables in the form of cookies will be needed to remember this user setting.